requirements.txt
numpy
langchain-core==0.3.67
openai
pydantic
pydantic-settings
pydantic_core
mcp
langchain
Flask
bcrypt
pandas
python-docx 
webvtt-py
azure-storage-blob
Flask-session
werkzeug
requests
openpyxl
tabulate
pandasql
matplotlib
html2text
beautifulsoup4
flask-cors
opencensus
opencensus-ext-flask
opencensus-ext-azure
identity
python-dotenv
azure-search-documents
azure-ai-formrecognizer
azure-cognitiveservices-speech
azure-identity
msgraph-sdk
pydub
moviepy
tiktoken
Markdown
weasyprint
PyYAML
openai-agents
# openai-agents
streamlit
openai